// asset/src/common/enums/relation-direction.enum.ts
export enum RelationDirection {
  DIRECTIONAL = 'directional',
  BIDIRECTIONAL = 'bidirectional',
}

export enum RelationDirectionType {
  HAS = 'Has',
  IS_INSTALLED_ON = 'Is Installed On',
  CONTAINS = 'Contains',
  BELONGS_TO = 'Belongs To',
  HOSTS = 'Hosts',
  RUNS = 'Runs',
  RELATED = 'Related',
  INSTALLED = 'Installed',
  INCLUDED = 'Included',
  CONTAINING = 'Containing',
  OWNED = 'Owned',
  REGISTERED = 'Registered',
}